+++ danra_to_zarr/pipeline/config.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from .. import __version__
 
 # FP_ROOT = Path("/dmidata/projects/cloudphysics/danra/data")
-FP_ROOT = Path("/nwp/danra/data")
+FP_ROOT = Path("/vf/danra/data")
 FP_TEMP_ROOT = Path("/nwp/danra/tempfiles")
 
 VERSION = f"v{__version__.split('+')[0]}"
@@ -15,18 +15,19 @@ DELETE_INTERMEDIATE_ZARR_FILES = True
 # Each part may contain multiple "level types" (e.g. heightAboveGround, etc)
 # and a name-mapping may also be defined
 
+t_start = "1991-01-01"
+t_end = "1992-01-02"
 
 DATA_COLLECTION = dict(
-    description="All prognostic variables for 2019-01-01 to 2024-01-01 reduced levels",
-    rechunk_to=dict(time=256, x=256, y=256),
-    intermediate_time_partitioning=["P14D", "P26W"],
-    timespan=slice("2019-01-01", "2024-01-01"),
+    description=f"All prognostic variables for {t_start} to {t_end} on all levels",
+    rechunk_to=dict(time=256, x=256, y=256, pressure=1, altitude=1),
+    intermediate_time_partitioning=["P1D", "P2W", "P26W"],
+    timespan=slice(t_start, t_end),
     parts=dict(
         height_levels=dict(
             heightAboveGround=dict(
                 variables={
-                    # v: [30, 50, 75, 100, 150, 200, 250, 300, 500]
-                    v: [100]
+                    v: [30, 50, 75, 100, 150, 200, 250, 300, 500]
                     for v in "t r u v".split()
                 }
             )
@@ -34,83 +35,97 @@ DATA_COLLECTION = dict(
         pressure_levels=dict(
             isobaricInhPa=dict(
                 variables={
-                    # v: [ 100, 200, 250, 300, 400, 500, 600, 700, 800, 850, 900, 925, 950, 1000, ]
-                    v: [1000]
+                    v: [
+                        1000,
+                        950,
+                        925,
+                        900,
+                        850,
+                        800,
+                        700,
+                        600,
+                        500,
+                        400,
+                        300,
+                        250,
+                        200,
+                        100,
+                    ]
                     for v in "z t u v tw r ciwc cwat".split()
                 }
             )
         ),
-        single_levels=dict(
-            heightAboveGround=dict(
-                variables={
-                    **{
-                        v: [0]
-                        for v in [
-                            "hcc",
-                            "icei",
-                            "lcc",
-                            "lwavr",
-                            "mcc",
-                            "mld",
-                            "pres",
-                            "prtp",
-                            "psct",
-                            "pscw",
-                            "pstb",
-                            "pstbc",
-                            "sf",
-                            "swavr",
-                            "vis",
-                            "xhail",
-                            # TODO: `rain`, `snow`, `snsub`, `nlwrs`, `nswrs`, `lhsub`,
-                            # `lhe`, `grpl`, `dni`, `grad`, `sshf`, `wevap`, `vflx`, `uflx`, `tp`
-                            # use time-range indicator 4, but dmidc currently assumes 0
-                            # "rain",
-                            # "snow",
-                            # "snsub",
-                            # "nlwrs",
-                            # "nswrs",
-                            # "lhsub",
-                            # "lhe",
-                            # "grpl",
-                            # "dni",
-                            # "grad",
-                            # "sshf",
-                            # "wevap",
-                            # "vflx",
-                            # "uflx",
-                            # "tp",
-                        ]
-                    },
-                    **{
-                        "t": [0, 2],
-                        # TODO: `tmin` and `tmax` use time-range indicator 2, but dmidc currently assumes 0
-                        # "tmin": [2],
-                        # "tmax": [2],
-                        "r": [2],
-                        "u": [10],
-                        "v": [10],
-                        # TODO: in next version add ugst and vgst, but dmidc needs to be updated first
-                        # to change timerange_indicator to 10 (rather than 0 by default), need to find
-                        # out what value 10 means too..
-                        # "ugst": [10],
-                        # "vgst": [10],
-                    },
-                },
-                level_name_mapping="{var_name}{level}m",
-            ),
-            entireAtmosphere=dict(
-                variables={v: [0] for v in "pwat cape cb ct grpl".split()},
-                level_name_mapping="{var_name}_column",
-            ),
-            # the variables `nswrt` and `nlwrt` are not available with timeRangeIndicator==0
-            # for all source tar-files, and so I am excluding them for now
-            # nominalTop=dict(
-            #     variables=dict(nswrt=[0], nlwrt=[0]), level_name_mapping="{var_name}_toa"
-            # ),
-            heightAboveSea=dict(
-                variables=dict(pres=[0]), level_name_mapping="{var_name}_seasurface"
-            ),
-        ),
+        # single_levels=dict(
+        #     heightAboveGround=dict(
+        #         variables={
+        #             **{
+        #                 v: [0]
+        #                 for v in [
+        #                     "hcc",
+        #                     "icei",
+        #                     "lcc",
+        #                     "lwavr",
+        #                     "mcc",
+        #                     "mld",
+        #                     "pres",
+        #                     "prtp",
+        #                     "psct",
+        #                     "pscw",
+        #                     "pstb",
+        #                     "pstbc",
+        #                     "sf",
+        #                     "swavr",
+        #                     "vis",
+        #                     "xhail",
+        #                     # TODO: `rain`, `snow`, `snsub`, `nlwrs`, `nswrs`, `lhsub`,
+        #                     # `lhe`, `grpl`, `dni`, `grad`, `sshf`, `wevap`, `vflx`, `uflx`, `tp`
+        #                     # use time-range indicator 4, but dmidc currently assumes 0
+        #                     # "rain",
+        #                     # "snow",
+        #                     # "snsub",
+        #                     # "nlwrs",
+        #                     # "nswrs",
+        #                     # "lhsub",
+        #                     # "lhe",
+        #                     # "grpl",
+        #                     # "dni",
+        #                     # "grad",
+        #                     # "sshf",
+        #                     # "wevap",
+        #                     # "vflx",
+        #                     # "uflx",
+        #                     # "tp",
+        #                 ]
+        #             },
+        #             **{
+        #                 "t": [0, 2],
+        #                 # TODO: `tmin` and `tmax` use time-range indicator 2, but dmidc currently assumes 0
+        #                 # "tmin": [2],
+        #                 # "tmax": [2],
+        #                 "r": [2],
+        #                 "u": [10],
+        #                 "v": [10],
+        #                 # TODO: in next version add ugst and vgst, but dmidc needs to be updated first
+        #                 # to change timerange_indicator to 10 (rather than 0 by default), need to find
+        #                 # out what value 10 means too..
+        #                 # "ugst": [10],
+        #                 # "vgst": [10],
+        #             },
+        #         },
+        #         level_name_mapping="{var_name}{level}m",
+        #     ),
+        #     entireAtmosphere=dict(
+        #         variables={v: [0] for v in "pwat cape cb ct grpl".split()},
+        #         level_name_mapping="{var_name}_column",
+        #     ),
+        #     # the variables `nswrt` and `nlwrt` are not available with timeRangeIndicator==0
+        #     # for all source tar-files, and so I am excluding them for now
+        #     # nominalTop=dict(
+        #     #     variables=dict(nswrt=[0], nlwrt=[0]), level_name_mapping="{var_name}_toa"
+        #     # ),
+        #     heightAboveSea=dict(
+        #         variables=dict(pres=[0]), level_name_mapping="{var_name}_seasurface"
+        #     ),
+        # ),
     ),
 )